Moved the single account operations to the account module
