+++ src/account/mod.rs
@@ -7,6 +7,51 @@ use r2d2_sqlite::SqliteConnectionManager;
 pub mod data;
 mod db;
 
+use crate::datastruct;
+
+pub fn get_account(
+    params: web::Path<datastruct::IdRequest>,
+    pool: web::Data<Pool<SqliteConnectionManager>>,
+) -> impl Future<Item = HttpResponse, Error = Error> {
+    let conn = pool.get().unwrap();
+    let result = db::get_account(conn, params.id);
+
+    match result {
+        Ok(v) => ok(HttpResponse::Ok().json(v)),
+        Err(_e) => ok(HttpResponse::InternalServerError().finish()),
+    }
+}
+
+pub fn create_account(
+    account: web::Json<data::NewAccount>,
+    pool: web::Data<Pool<SqliteConnectionManager>>,
+) -> impl Future<Item = HttpResponse, Error = Error> {
+    let account_type = data::AccountType::from_i32(account.acc_type);
+    let result = db::add_account(
+        pool.get().unwrap(),
+        account_type,
+        &account.name,
+        &account.currency,
+    );
+
+    match result {
+        Ok(v) => ok(HttpResponse::Ok().json(v)),
+        Err(_e) => ok(HttpResponse::InternalServerError().finish()),
+    }
+}
+
+pub fn delete_account(
+    params: web::Path<datastruct::IdRequest>,
+    pool: web::Data<Pool<SqliteConnectionManager>>,
+) -> impl Future<Item = HttpResponse, Error = Error> {
+    let result = db::remove_account(pool.get().unwrap(), params.id);
+
+    match result {
+        Ok(_v) => ok(HttpResponse::Ok().finish()),
+        Err(_e) => ok(HttpResponse::InternalServerError().finish()),
+    }
+}
+
 pub fn list_accounts(
     pool: web::Data<Pool<SqliteConnectionManager>>,
 ) -> impl Future<Item = HttpResponse, Error = Error> {
